❌ Breaking Change: Implement the new Pause State from the rive timer animation.
The original pause state "Stops" the animation with a confetti, but I modified the rive animation to include a pause state and trigger which is basically the "Stops" the animation without a confetti. With this change, I could simply add a stop trigger to handle the original Stop animation which was using the "Pause" trigger in this app. But not it gets it's own trigger and the Pause trigger I had locally can continue handling but this time handle the new Pause state